NLP Service (HuggingFace + FAISS)/requirements.txt
flask==3.0.2
sentence-transformers==2.6.1
faiss-cpu==1.8.0.post1
numpy==1.26.4

